+++ CHANGELOG.rst
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on `Keep a Changelog <http://keepachangelog.com/>`_.
 
 =====================
-22.10.04 - 22.11.04
+22.10.04 - 22.10.04
 =====================
 * Add Compute Instance Plugin Status to the JSON output and CSV